requirements.txt
aiogram>=3.13.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
pydantic>=2.9.2
pydantic-settings>=2.5.2
python-dotenv>=1.0.1
httpx>=0.27.2
openai>=1.51.0
python-dateutil>=2.9.0.post0
google-api-python-client>=2.149.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
apscheduler>=3.10.4
